+++ NEWS.md
@@ -1,3 +1,17 @@
+## version 3.3
+### Features
+- the spike and slab algorithm for faster model estimation via the `algorithm = "ss"` argument (see a new vignette for more details)
+- refactoring of the JAGS C++ code of weighted distributions and exporting of the lpdfs into JAGS (maintenance)
+- weights_mix JAGS prior distribution to sample a mixture of weight functions directly
+
+### Fixes
+- incorrectly omitting models with more than one predictor when computing conditional marginal summary
+
+## version 3.2.1
+### Features
+- default prior distributions for all parameters are now set via the `set_default_priors()` function
+- `rescale_priors` argument allows to conveniently re-scale the prior distributions for the effect, heterogeneity, and bias simultaneously
+
 ## version 3.2
 ### Features
 - `summary_heterogeneity()` function to summarize the heterogeneity of the RoBMA models (prediction interval, tau, tau^2, I^2, and H^2)